sanphelp_microservice/user_application/requirements.txt
amqp==1.4.8
anyjson==0.3.3
awscli==1.11.129
boto==2.38.0
boto3==1.4.4
botocore==1.5.92
Django==1.8.4
django-braces==1.8.1
django-cors-headers==1.1.0
django-extensions==1.5.7
django-filter==0.11.0
django-rest-swagger==0.3.5
django-storages==1.1.8
djangorestframework==3.2.4
enum34==1.0.4
drfdocs==0.0.6
Markdown==2.6.2
meld3==1.0.2
mysqlclient==1.3.10
MySQL-python==1.2.5
Pillow==2.9.0
pyparsing==1.5.7
pytz==2015.7
requests==2.7.0
six==1.9.0
srp==1.0.7
validate-email==1.3


